DCS-778 Find locations satisfying an Video Link Appointment Locations Specification.
