+++ src/main/java/uk/gov/justice/digital/hmpps/whereabouts/services/locationfinder/AppointmentLocationsService.kt
@@ -17,7 +17,12 @@ class AppointmentLocationsService(
       .filterNot { excludedAppointmentIds.contains(it.id) }
 
     val locationsForAppointmentIntervals =
-      appointmentLocationsFinderService.find(specification, allLocations, scheduledAppointments)
+      appointmentLocationsFinderService.find(
+        specification.appointmentIntervals,
+        allLocations.map { it.locationId },
+        scheduledAppointments
+      )
+
     return toAvailableLocations(locationsForAppointmentIntervals, allLocations)
   }
